WIP: admin vulnerability metrics progress
