+++ docs/VULNERABILITY_DASHBOARD_CLEANUP_SUMMARY.md
@@ -1,0 +1,70 @@
+# Vulnerability Dashboard Cleanup Summary
+
+Audience: reviewers preparing the admin vulnerability dashboard for a pull request.
+
+## Scope
+
+This cleanup pass stayed inside the admin vulnerability dashboard and reviewer-support surface:
+
+- `project/application/vulnerability_mgmt/admin.py`
+- `project/application/templates/admin/vulnerability_metrics.html`
+- `project/application/templates/admin/vulnerability_host_detail.html`
+- `project/application/vulnerability_mgmt/tests.py`
+- `project/application/main/management/commands/import_dev_scan_logs.py`
+- `docs/VULNERABILITY_METRICS_ARCHITECTURE.md`
+- `docs/VULNERABILITY_METRICS_FLOW.md`
+
+Scanner lifecycle and broader host-management code were left out of scope unless they directly affected dashboard reviewability.
+
+## What Was Improved
+
+### Backend readability
+
+- Added named constants for the default dashboard range, severity colors, severity chart dimensions, and trend chart dimensions.
+- Extracted shared date-range parsing into `_selected_date_range()`.
+- Extracted timezone-aware date window construction into `_datetime_bounds()`.
+- Added `_metrics_query_string()` so dashboard links preserve state through one helper instead of repeated template strings.
+- Added QuerySet type hints to the deduplication and mitigation query helpers.
+- Expanded the deduplication docstring to explain why the dashboard uses the highest primary key as the representative row.
+- Removed the unused `trend_hover_rows` accumulator from the trend hover-info construction.
+
+### Bug and resilience cleanup
+
+- Made scan-mode mitigation event selection tolerate events that do not contain a `mitigated_at` value.
+- Kept the existing friendly admin error behavior for missing tables and unexpected dashboard failures.
+- Removed unused logging setup from the optional `import_dev_scan_logs` command.
+
+### Template cleanup
+
+- Replaced repeated hand-built query strings in the metrics template with context-provided query values.
+- Kept the template behavior the same for date presets, mode toggles, host sorting, trend sorting, and host detail links.
+- Wrapped touched host navigation links for readability.
+
+### Tests
+
+- Added coverage for dashboard query-string generation.
+- Added coverage for scan-mode mitigation selection when a malformed event lacks `mitigated_at`.
+- Existing mitigation tests still cover same-task comparison, different-task isolation, and reappearing findings in persisted mode.
+
+### Documentation
+
+- Removed stale `avg_cvss` references from dashboard docs.
+- Updated trend documentation from a fixed 14-day window to the selected date window.
+- Updated documented aggregation tools to match the current implementation.
+- Clarified that critical and high counts are the severity KPI values currently rendered.
+
+## Behavior Notes
+
+The intended dashboard behavior remains the same:
+
+- Metrics are still calculated at request time from `Vulnerability` rows.
+- The dashboard still supports scan mode, persisted mode, distinct mode, date presets, host sorting, trend sorting, and host drill-down links.
+- The admin navigation patch remains guarded against duplicate URL/app-list registration.
+- The optional dev importer still creates `ScanReport` placeholders for imported scan-result files.
+
+## Validation Done So Far
+
+- `python3 -m py_compile` passes for the edited Python files.
+- VS Code reports no template errors for the dashboard templates.
+- Local editor diagnostics for `admin.py` still show missing Django imports because the local analysis environment does not have Django installed.
+- The local virtualenv at `/root/deterrers-venv` does not include Django, so full Django validation should run in the dev container after rebuilding the image from the edited source.
