+++ docs/VULNERABILITY_DASHBOARD_REVIEW_ISSUES.md
@@ -1,0 +1,51 @@
+# Vulnerability Dashboard Review Issues
+
+Audience: reviewer checklist for deciding whether the admin vulnerability dashboard is ready for a pull request.
+
+## Current Blockers
+
+No source-level blocker is known from this cleanup pass, but full Django validation still needs to run against a rebuilt dev container so it exercises the edited source files.
+
+## Must Review Before PR
+
+1. Container validation
+
+   The running `project-dev-web-1` container may not mount source code live. Rebuild the dev web image before treating `manage.py check`, dashboard render smoke tests, or Django unit tests as final validation.
+
+2. Dashboard render smoke test
+
+   Confirm that `/admin/vulnerability-metrics/` renders after rebuild and that these interactions preserve state:
+
+   - date range apply
+   - scan/persist mode toggle
+   - distinct mode toggle
+   - host sorting
+   - trend sorting
+   - top-host drill-down and back link
+
+3. Mitigation semantics
+
+   The dashboard counts a mitigation when a finding disappears from a newer comparable report for the same host and scanner task. Review whether `host_ipv4 + task_id` is the correct comparison boundary for manual, registration, and periodic scans.
+
+4. Distinct-mode representative row
+
+   `_dedupe_findings_queryset()` uses the highest primary key as the representative row for each host/service/finding identity. Review whether this is acceptable for dashboard counts or whether the representative should be selected by timestamp/report ordering instead.
+
+5. Performance of persisted trend mode
+
+   Persisted mode still performs per-day aggregate queries over the snapshot range. This is acceptable for small review data but should be benchmarked against large vulnerability tables before production rollout.
+
+## Non-Blocking Follow-Ups
+
+- Consider replacing broad `except Exception` handlers in the admin views with narrower failure handling after the dashboard has stronger test coverage.
+- Decide whether the Flatpickr CDN dependency should be vendored into project static files for offline or restricted-network deployments.
+- Add render-level tests for the metrics and host detail templates once a stable Django test database setup is available.
+- Harden `import_dev_scan_logs` numeric conversion if future dev logs may contain null or malformed `qod`, `cvss_version`, or `cvss_base_score` fields.
+- Review timezone/DST behavior for ranges that cross daylight-saving transitions, especially because the project timezone is `Europe/Berlin`.
+- Consider moving larger dashboard computation pieces out of `admin.py` if future changes grow the feature beyond reviewable admin integration logic.
+
+## Reviewer Decision Points
+
+- Is request-time computation still preferred over materialized dashboard metrics for the expected data volume?
+- Should the PR include the optional dev importer, or should it remain a separate reviewer-support change?
+- Are the dashboard docs under `docs/` intended to be committed with the feature PR, given that the docs directory is currently untracked in this worktree?
