+++ docs/VULNERABILITY_METRICS_ARCHITECTURE.md
@@ -1,0 +1,139 @@
+# Vulnerability Metrics Architecture Note
+
+Audience: contributors and reviewers working on the admin metrics feature.
+
+## 1) Feature Goal
+
+Provide a dedicated Django Admin page that gives a fast vulnerability overview:
+- summary KPIs
+- severity distribution
+- trend visualization
+- top affected hosts
+- top findings (NVT names)
+
+The page should work with real scan data and with imported development data.
+
+## 2) Core Files (Feature Boundary)
+
+These files are the minimal feature surface for admin metrics:
+
+1. `application/vulnerability_mgmt/admin.py`
+- registers custom admin URL `admin/vulnerability-metrics/`
+- injects a virtual "Vulnerability Metrics" app into the admin app list
+- computes all page metrics and SVG chart data from `Vulnerability` rows
+
+2. `application/templates/admin/index.html`
+- keeps Django default app list rendering (`{{ block.super }}`)
+- relies on the app-list injection from `admin.py` so metrics appears in left navigation
+
+3. `application/templates/admin/vulnerability_metrics.html`
+- renders KPI cards, tables, and inline SVG charts
+- consumes precomputed context from `vulnerability_metrics_view`
+
+## 3) Data Source and Storage
+
+Primary source for this page:
+- model: `vulnerability_mgmt.Vulnerability`
+- database alias: `vulnerability_mgmt`
+
+Important detail:
+- the metrics page is aggregation-only; it does not maintain precomputed metric tables
+- all values are calculated at request time
+
+Related but secondary:
+- `vulnerability_mgmt.ScanReport` stores XML reports
+- not required for rendering metrics cards/charts directly
+
+## 4) Request/Data Flow
+
+```text
+Admin dashboard load
+  -> admin.site.get_app_list() patch appends virtual app section
+  -> user clicks "Vulnerability Metrics / Overview"
+  -> custom admin URL routes to vulnerability_metrics_view
+  -> ORM queries on Vulnerability (using('vulnerability_mgmt'))
+  -> view computes counters, rankings, trend, SVG coordinates
+  -> template renders cards + tables + inline SVG charts
+```
+
+## 5) Metric Computation Summary
+
+Computed in `vulnerability_metrics_view`:
+- total, active, silenced, and mitigated vulnerability counts
+- mitigated counts are computed by comparing consecutive reports for the same host and scanner task; a finding is mitigated when its stable identity (`nvt_oid`/name + port + protocol) existed in an older report and is absent from a newer comparable report
+- distinct host count
+- distinct report count
+- latest detection timestamp
+- severity buckets (critical/high/medium/low/unknown)
+- critical and high finding counts for KPI cards
+- top hosts by finding count (with active/silenced/mitigated/max CVSS)
+- top findings by NVT name
+- selected trend window (total + active, with silenced and mitigated values in the table)
+
+Chart preparation:
+- severity stacked bar segments (`x`, `width`, `color`)
+- trend line coordinates (`total_points`, `active_points`)
+- y-grid and x-label positions for SVG axes
+
+## 6) Optional Supporting File (Data Seeding)
+
+`application/main/management/commands/import_dev_scan_logs.py`
+- imports `scan-results_*.json` into `Vulnerability`
+- now also creates `ScanReport` placeholders per imported report id
+- useful for local testing/review when scanner queue is slow
+
+This command is recommended for reviewer reproducibility, but it is not strictly required to render the metrics feature if data already exists.
+
+## 7) Out-of-Scope for Core Metrics PR
+
+These files are scanner lifecycle/reliability concerns, not core admin metrics rendering:
+- `application/main/views.py`
+- `application/main/management/commands/sync_scanner.py`
+- `application/main/core/scanner/scanner_mock.py`
+
+They can be delivered in a separate PR to keep review scope clear.
+
+## 8) Reviewer Checklist
+
+1. Navigation
+- metrics entry visible in left admin app list
+- link opens `/admin/vulnerability-metrics/`
+
+2. Data correctness
+- counters match DB values in `Vulnerability`
+- trend and severity values are internally consistent
+
+3. Resilience
+- page shows friendly `db_error` if vulnerability tables are missing
+
+4. Performance sanity
+- no N+1 loops over individual rows
+- all main sections use aggregate queries/annotations
+
+5. UI integrity
+- SVG charts render without JS dependencies
+- tables remain readable for small and large datasets
+
+## 9) Quick Validation Commands
+
+```bash
+# Django integrity
+cd /root/deterrers/project
+docker exec project-dev-web-1 python manage.py check
+
+# Render view in shell (smoke test)
+docker exec project-dev-web-1 python manage.py shell -c "from django.test import RequestFactory; from django.contrib.auth import get_user_model; from vulnerability_mgmt.admin import vulnerability_metrics_view; User=get_user_model(); u=User.objects.get(username='deterrers-admin'); req=RequestFactory().get('/admin/vulnerability-metrics/', HTTP_HOST='localhost'); req.user=u; resp=vulnerability_metrics_view(req); resp.render(); print(resp.status_code)"
+```
+
+## 10) Recommended PR Split Strategy
+
+PR A (metrics feature):
+- `application/vulnerability_mgmt/admin.py`
+- `application/templates/admin/index.html`
+- `application/templates/admin/vulnerability_metrics.html`
+- optionally `application/main/management/commands/import_dev_scan_logs.py` for test data reproducibility
+
+PR B (scanner reliability):
+- `application/main/views.py`
+- `application/main/management/commands/sync_scanner.py`
+- `application/main/core/scanner/scanner_mock.py`
