+++ project/application/templates/admin/vulnerability_latest_host_detail.html
@@ -1,0 +1,151 @@
+{% extends "admin/base_site.html" %}
+{% load i18n %}
+
+{% block extrastyle %}
+{{ block.super }}
+<style>
+  .latest-detail-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(190px, 1fr));
+    gap: 12px;
+    margin-bottom: 16px;
+  }
+
+  .latest-detail-card {
+    border: 1px solid var(--hairline-color);
+    border-radius: 4px;
+    background: var(--body-bg);
+    padding: 12px;
+  }
+
+  .latest-detail-title {
+    font-size: 12px;
+    text-transform: uppercase;
+    color: var(--body-quiet-color);
+    margin-bottom: 8px;
+  }
+
+  .latest-detail-value {
+    font-size: 24px;
+    line-height: 1.1;
+    font-weight: 600;
+  }
+
+  .latest-back-link {
+    display: inline-block;
+    margin-bottom: 12px;
+  }
+
+  .finding-name {
+    max-width: 520px;
+    word-break: break-word;
+  }
+
+  .finding-meta {
+    margin-top: 3px;
+    font-size: 11px;
+    color: var(--body-quiet-color);
+  }
+</style>
+{% endblock %}
+
+{% block breadcrumbs %}
+<div class="breadcrumbs">
+  <a href="{% url 'admin:index' %}">{% translate "Home" %}</a>
+  &rsaquo; <a href="{% url 'admin:vulnerability_metrics' %}">{% translate "Vulnerability Metrics" %}</a>
+  &rsaquo; <a href="{% url 'admin:vulnerability_latest_reports' %}">{% translate "Latest reports" %}</a>
+  &rsaquo; {% translate "Host Details" %}
+</div>
+{% endblock %}
+
+{% block content %}
+<div id="content-main">
+  {% if db_error %}
+  <div class="errornote">{{ db_error }}</div>
+  {% else %}
+  <a
+    class="latest-back-link"
+    href="{% url 'admin:vulnerability_latest_reports' %}"
+  >&larr; {% translate "Back to latest reports" %}</a>
+
+  <h1>{{ host_ipv4 }}</h1>
+  <p>{% translate "Latest state built from the newest report per scanner task for this host." %}</p>
+
+  <div class="latest-detail-grid">
+    <div class="latest-detail-card">
+      <div class="latest-detail-title">{% translate "Total" %}</div>
+      <div class="latest-detail-value">{{ host_summary.total }}</div>
+    </div>
+    <div class="latest-detail-card">
+      <div class="latest-detail-title">{% translate "Active" %}</div>
+      <div class="latest-detail-value">{{ host_summary.active }}</div>
+    </div>
+    <div class="latest-detail-card">
+      <div class="latest-detail-title">{% translate "Silenced" %}</div>
+      <div class="latest-detail-value">{{ host_summary.silenced }}</div>
+    </div>
+    <div class="latest-detail-card">
+      <div class="latest-detail-title">{% translate "Critical" %}</div>
+      <div class="latest-detail-value">{{ host_summary.critical }}</div>
+    </div>
+    <div class="latest-detail-card">
+      <div class="latest-detail-title">{% translate "High" %}</div>
+      <div class="latest-detail-value">{{ host_summary.high }}</div>
+    </div>
+    <div class="latest-detail-card">
+      <div class="latest-detail-title">{% translate "Max CVSS" %}</div>
+      <div class="latest-detail-value">{{ host_summary.max_cvss|default:"-" }}</div>
+    </div>
+    <div class="latest-detail-card">
+      <div class="latest-detail-title">{% translate "Tasks" %}</div>
+      <div class="latest-detail-value">{{ host_summary.tasks }}</div>
+    </div>
+    <div class="latest-detail-card">
+      <div class="latest-detail-title">{% translate "Reports" %}</div>
+      <div class="latest-detail-value">{{ host_summary.reports }}</div>
+    </div>
+  </div>
+
+  <div class="module">
+    <h2>{% translate "Latest findings" %}</h2>
+    <table>
+      <thead>
+        <tr>
+          <th>{% translate "Detected" %}</th>
+          <th>{% translate "Finding" %}</th>
+          <th>{% translate "OID" %}</th>
+          <th>{% translate "CVSS" %}</th>
+          <th>{% translate "Status" %}</th>
+          <th>{% translate "Port/Proto" %}</th>
+          <th>{% translate "Task ID" %}</th>
+          <th>{% translate "Report ID" %}</th>
+          <th>{% translate "UUID" %}</th>
+        </tr>
+      </thead>
+      <tbody>
+        {% for row in latest_findings %}
+        <tr>
+          <td>{{ row.date_time|date:"Y-m-d H:i" }}</td>
+          <td class="finding-name">
+            {{ row.nvt_name|default:row.vulnerability_name|default:"-" }}
+            <div class="finding-meta">{{ row.vulnerability_name|default:"" }}</div>
+          </td>
+          <td>{{ row.nvt_oid|default:"-" }}</td>
+          <td>{{ row.cvss_base_score|default:"-" }}</td>
+          <td>{% if row.is_silenced %}{% translate "Silenced" %}{% else %}{% translate "Active" %}{% endif %}</td>
+          <td>{{ row.port|default:"-" }}/{{ row.proto|default:"-" }}</td>
+          <td>{{ row.task_id }}</td>
+          <td>{{ row.report_id }}</td>
+          <td>{{ row.uuid|default:"-" }}</td>
+        </tr>
+        {% empty %}
+        <tr>
+          <td colspan="9">{% translate "No latest findings found for this host." %}</td>
+        </tr>
+        {% endfor %}
+      </tbody>
+    </table>
+  </div>
+  {% endif %}
+</div>
+{% endblock %}
