+++ project/application/templates/admin/vulnerability_latest_reports.html
@@ -1,0 +1,164 @@
+{% extends "admin/base_site.html" %}
+{% load i18n %}
+
+{% block extrastyle %}
+{{ block.super }}
+<style>
+  .latest-grid {
+    display: grid;
+    grid-template-columns: repeat(auto-fit, minmax(190px, 1fr));
+    gap: 12px;
+    margin-bottom: 16px;
+  }
+
+  .latest-card {
+    border: 1px solid var(--hairline-color);
+    border-radius: 4px;
+    background: var(--body-bg);
+    padding: 12px;
+  }
+
+  .latest-card-title {
+    font-size: 12px;
+    text-transform: uppercase;
+    color: var(--body-quiet-color);
+    margin-bottom: 8px;
+  }
+
+  .latest-card-value {
+    font-size: 24px;
+    line-height: 1.1;
+    font-weight: 600;
+  }
+
+  .latest-note {
+    color: var(--body-quiet-color);
+    margin-bottom: 16px;
+  }
+
+  .host-link {
+    color: var(--link-fg);
+    text-decoration: none;
+  }
+
+  .host-link:hover,
+  .host-link:focus {
+    text-decoration: underline;
+  }
+
+  .bar-cell {
+    min-width: 180px;
+  }
+
+  .mini-bar {
+    width: 100%;
+    height: 12px;
+    display: block;
+  }
+</style>
+{% endblock %}
+
+{% block breadcrumbs %}
+<div class="breadcrumbs">
+  <a href="{% url 'admin:index' %}">{% translate "Home" %}</a>
+  &rsaquo; <a href="{% url 'admin:vulnerability_metrics' %}">{% translate "Vulnerability Metrics" %}</a>
+  &rsaquo; {% translate "Latest reports" %}
+</div>
+{% endblock %}
+
+{% block content %}
+<div id="content-main">
+  {% if db_error %}
+  <div class="errornote">{{ db_error }}</div>
+  {% else %}
+  <p class="latest-note">
+    {% translate "Current latest state built from the newest report per host and scanner task." %}
+  </p>
+
+  <div class="latest-grid">
+    <div class="latest-card">
+      <div class="latest-card-title">{% translate "Affected hosts" %}</div>
+      <div class="latest-card-value">{{ summary.hosts }}</div>
+    </div>
+    <div class="latest-card">
+      <div class="latest-card-title">{% translate "Latest findings" %}</div>
+      <div class="latest-card-value">{{ summary.total }}</div>
+    </div>
+    <div class="latest-card">
+      <div class="latest-card-title">{% translate "Active" %}</div>
+      <div class="latest-card-value">{{ summary.active }}</div>
+    </div>
+    <div class="latest-card">
+      <div class="latest-card-title">{% translate "Silenced" %}</div>
+      <div class="latest-card-value">{{ summary.silenced }}</div>
+    </div>
+    <div class="latest-card">
+      <div class="latest-card-title">{% translate "Critical" %}</div>
+      <div class="latest-card-value">{{ summary.critical }}</div>
+    </div>
+    <div class="latest-card">
+      <div class="latest-card-title">{% translate "High" %}</div>
+      <div class="latest-card-value">{{ summary.high }}</div>
+    </div>
+    <div class="latest-card">
+      <div class="latest-card-title">{% translate "Tasks" %}</div>
+      <div class="latest-card-value">{{ summary.tasks }}</div>
+    </div>
+    <div class="latest-card">
+      <div class="latest-card-title">{% translate "Reports" %}</div>
+      <div class="latest-card-value">{{ summary.reports }}</div>
+    </div>
+  </div>
+
+  <div class="module">
+    <h2>{% translate "Affected hosts in latest reports" %}</h2>
+    <table>
+      <thead>
+        <tr>
+          <th>{% translate "Host IPv4" %}</th>
+          <th>{% translate "Active" %}</th>
+          <th>{% translate "Silenced" %}</th>
+          <th>{% translate "Critical" %}</th>
+          <th>{% translate "High" %}</th>
+          <th>{% translate "Max CVSS" %}</th>
+          <th>{% translate "Tasks" %}</th>
+          <th>{% translate "Reports" %}</th>
+          <th>{% translate "Latest report" %}</th>
+          <th>{% translate "Load" %}</th>
+        </tr>
+      </thead>
+      <tbody>
+        {% for host in latest_hosts %}
+        <tr>
+          <td>
+            <a
+              class="host-link"
+              href="{% url 'admin:vulnerability_latest_host_detail' host.host_ipv4 %}"
+            >{{ host.host_ipv4 }}</a>
+          </td>
+          <td>{{ host.active }}</td>
+          <td>{{ host.silenced }}</td>
+          <td>{{ host.critical }}</td>
+          <td>{{ host.high }}</td>
+          <td>{{ host.max_cvss|default:"-" }}</td>
+          <td>{{ host.tasks }}</td>
+          <td>{{ host.reports }}</td>
+          <td>{{ host.latest_report_timestamp|date:"Y-m-d H:i"|default:"-" }}</td>
+          <td class="bar-cell">
+            <svg class="mini-bar" viewBox="0 0 100 10" preserveAspectRatio="none">
+              <rect x="0" y="0" width="100" height="10" fill="#283741"></rect>
+              <rect x="0" y="0" width="{{ host.pct }}" height="10" fill="{{ host.bar_color }}"></rect>
+            </svg>
+          </td>
+        </tr>
+        {% empty %}
+        <tr>
+          <td colspan="10">{% translate "No affected hosts found in latest reports." %}</td>
+        </tr>
+        {% endfor %}
+      </tbody>
+    </table>
+  </div>
+  {% endif %}
+</div>
+{% endblock %}
