+++ project/application/templates/admin/vulnerability_metrics_scratch.html
@@ -1,0 +1,85 @@
+{% extends "admin/base_site.html" %}
+{% load i18n %}
+
+{% block breadcrumbs %}
+<div class="breadcrumbs">
+  <a href="{% url 'admin:index' %}">{% translate "Home" %}</a>
+  &rsaquo; {% translate "Vulnerability Metrics" %}
+  &rsaquo; {{ page_title }}
+</div>
+{% endblock %}
+
+// Time range selection
+{% block content %}
+<div id="content-main">
+  <div style="border: 1px solid; border-radius: 8px; padding: 12px 14px; margin-bottom: 12px;">
+    <h1 style="margin: 0 0 6px 0;">{{ page_title }}</h1>
+    
+    <!-- Predefined time range dropdown -->
+    <form method="get" style="margin-bottom: 12px; font-size: 13px; display: flex; align-items: center; gap: 8px; flex-wrap: wrap;">
+      <label for="days-select">Quick select:</label>
+      <select id="days-select" name="days" onchange="this.form.submit()" style="padding: 2px 4px;">
+        <!--<option value="">-- Choose --</option>-->
+        <option value="7" {% if days == 7 %}selected{% endif %}>Last 7 days</option>
+        <option value="14" {% if days == 14 %}selected{% endif %}>Last 14 days</option>
+        <option value="30" {% if days == 30 %}selected{% endif %}>Last 30 days</option>
+        <option value="90" {% if days == 90 %}selected{% endif %}>Last 90 days</option>
+      </select>
+
+      <input type="text" id="start-date" name="start_date" placeholder="From (DD.MM.YYYY)" value="{{ start_date }}" style="padding: 2px 4px; width: 120px;">
+      <span>–</span>
+      <input type="text" id="end-date" name="end_date" placeholder="To (DD.MM.YYYY)" value="{{ end_date }}" style="padding: 2px 4px; width: 120px;">
+      <button type="submit" style="padding: 2px 8px; cursor: pointer;">Filter</button>
+
+      {% if time_range %}
+        <span style="font-size: 12px; opacity: 0.75; margin-left: 6px;">{{ time_range }}</span>
+      {% endif %}
+    </form>
+
+    <p style="margin: 0 0 8px 0; font-size: 15px;">
+      Total amount of vulnerabilities:
+      <strong style="font-size: 18px;">{{ metric_data.total_count }}</strong>
+    </p>
+    {% if metric_data %}
+      <p style="margin: 0; line-height: 1.9;">
+        <span style="display: inline-block; padding: 2px 8px; border-radius: 999px; border: 1px solid; margin-right: 6px;">Affected hosts: <strong>{{ metric_data.hosts_count }}</strong></span>
+        <span style="display: inline-block; padding: 2px 8px; border-radius: 999px; border: 1px solid; margin-right: 6px;">Critical: <strong>{{ metric_data.critical_count }}</strong></span>
+        <span style="display: inline-block; padding: 2px 8px; border-radius: 999px; border: 1px solid; margin-right: 6px;">High: <strong>{{ metric_data.high_count }}</strong></span>
+        <span style="display: inline-block; padding: 2px 8px; border-radius: 999px; border: 1px solid; margin-right: 6px;">Medium: <strong>{{ metric_data.medium_count }}</strong></span>
+        <span style="display: inline-block; padding: 2px 8px; border-radius: 999px; border: 1px solid; margin-right: 6px;">Low: <strong>{{ metric_data.low_count }}</strong></span>
+        <span style="display: inline-block; padding: 2px 8px; border-radius: 999px; border: 1px solid;">Mitigated: <strong>{{ metric_data.mitigated_count }}</strong></span>
+      </p>
+    {% endif %}
+  </div>
+
+  <details>
+    <summary><strong>Top affected hosts</strong></summary>
+
+    {% if top_affected_hosts %}
+      <table style="margin-top: 10px; border-collapse: collapse; min-width: 560px; width: 100%; max-width: 820px;">
+        <thead>
+          <tr>
+            <th style="text-align: left; padding: 5px 8px; border-bottom: 1px solid #ccc;">Host</th>
+            <th style="text-align: right; padding: 5px 8px; border-bottom: 1px solid #ccc;">Severity count</th>
+            <th style="text-align: right; padding: 5px 8px; border-bottom: 1px solid #ccc;">Vulnerability count</th>
+            <th style="text-align: right; padding: 5px 8px; border-bottom: 1px solid #ccc;">Max severity</th>
+          </tr>
+        </thead>
+        <tbody>
+          {% for host in top_affected_hosts %}
+            <tr>
+              <td style="padding: 5px 8px; border-bottom: 1px solid #eee;">{{ host.name }}</td>
+              <td style="padding: 5px 8px; border-bottom: 1px solid #eee; text-align: right;">{{ host.severity_count|floatformat:1 }}</td>
+              <td style="padding: 5px 8px; border-bottom: 1px solid #eee; text-align: right;">{{ host.vuln_count }}</td>
+              <td style="padding: 5px 8px; border-bottom: 1px solid #eee; text-align: right;">{{ host.max_severity|floatformat:1 }}</td>
+            </tr>
+          {% endfor %}
+        </tbody>
+      </table>
+    {% else %}
+      <p style="margin-top: 12px;">No host data available.</p>
+    {% endif %}
+  </details>
+  
+</div>
+{% endblock %}
