+++ project/application/vulnerability_mgmt/admin.py
@@ -1,1275 +1,101 @@
-"""
-Admin integration for the Vulnerability Metrics dashboard.
+"""Admin entrypoint for vulnerability metrics pages.
 
-This module intentionally keeps all metric calculations in one place and does
-not require extra DB tables:
-- It registers a custom admin page at /admin/vulnerability-metrics/.
-- It injects a virtual app entry so the page appears in the normal admin app
-    list on the left.
-- It computes all cards/charts from Vulnerability rows at request time.
+Keep this file focused on Django admin wiring. Views and metric/query logic
+should live in separate modules as the dashboard grows again.
 """
-
-from collections import Counter, defaultdict
-from datetime import date, datetime, time, timedelta
-import logging
-from urllib.parse import urlencode
 
 from django.contrib import admin
-from django.db import ProgrammingError, OperationalError
-from django.db.models import Count, Max, Q
-from django.db.models.query import QuerySet
-from django.template.response import TemplateResponse
 from django.urls import path, reverse
-from django.utils import timezone
 
-from vulnerability_mgmt.models import Vulnerability
-
-
-logger = logging.getLogger(__name__)
-
-# Shared dashboard defaults.
-DEFAULT_RANGE_DAYS = 14
-
-# Severity palette reused by cards, table bars, and SVG charts.
-COLOR_CRITICAL = '#d9534f'
-COLOR_HIGH = '#f08c00'
-COLOR_MEDIUM = '#f0c24b'
-COLOR_LOW = '#4fa7cd'
-COLOR_UNKNOWN = '#7e8d9b'
-COLOR_EMPTY_CHART = '#6f7f89'
-
-# SVG layout values are centralized so Python/template stay in sync.
-SEVERITY_CHART_WIDTH = 640
-SEVERITY_CHART_OFFSET = 18
-
-TREND_CHART_WIDTH = 760
-TREND_CHART_HEIGHT = 240
-TREND_CHART_LEFT_PAD = 46
-TREND_CHART_RIGHT_PAD = 16
-TREND_CHART_TOP_PAD = 12
-TREND_CHART_BOTTOM_PAD = 32
+from vulnerability_mgmt import admin_views
 
 
-def _parse_iso_date(value: str | None) -> date | None:
-    """Parse a YYYY-MM-DD string into a date, or return None."""
-    if not value:
-        return None
-    try:
-        return date.fromisoformat(value)
-    except ValueError:
-        return None
+# Internal app label used only for the virtual admin sidebar section.
+APP_LABEL = 'vulnerability_metrics'
 
 
-def _severity_tint(critical_count: int, high_count: int, total_count: int) -> str:
-    """Map critical/high density to a readable chart color."""
-    if total_count <= 0:
-        return COLOR_EMPTY_CHART
-
-    critical_share = critical_count / total_count
-    high_share = high_count / total_count
-
-    # High-only days should be orange, critical-heavy days should be red.
-    severity_score = (critical_share * 1.0) + (high_share * 0.65)
-    if critical_count > 0:
-        severity_score = min(1.0, severity_score + 0.2)
-    elif high_count > 0:
-        severity_score = min(0.8, severity_score)
-
-    if severity_score <= 0:
-        return COLOR_LOW
-    if severity_score < 0.45:
-        return COLOR_MEDIUM
-    if severity_score < 0.8:
-        return COLOR_HIGH
-    return COLOR_CRITICAL
-
-
-def _cvss_bar_color(cvss_value: float | None) -> str:
-    """Map CVSS to table bar colors for hosts/findings."""
-    if cvss_value is None:
-        return COLOR_UNKNOWN
-    if cvss_value >= 9.0:
-        return COLOR_CRITICAL
-    if cvss_value >= 7.0:
-        return COLOR_HIGH
-    if cvss_value >= 4.0:
-        return COLOR_MEDIUM
-    if cvss_value > 0:
-        return COLOR_LOW
-    return COLOR_UNKNOWN
-
-
-def _format_range_label(start_day: date, end_day: date, today: date) -> str:
-    """Return a human-readable label for the selected dashboard range."""
-    default_start = today - timedelta(days=DEFAULT_RANGE_DAYS - 1)
-    if start_day == end_day:
-        return start_day.strftime('%d.%m.%Y')
-    if start_day == default_start and end_day == today:
-        return 'Last 14 days'
-    return f"{start_day.strftime('%d.%m.%Y')} - {end_day.strftime('%d.%m.%Y')}"
-
-
-def _selected_date_range(request) -> tuple[date, date, date]:
-    """Read the requested date range and normalize reversed input."""
-    today = timezone.localdate()
-    default_start_day = today - timedelta(days=DEFAULT_RANGE_DAYS - 1)
-    selected_start_day = _parse_iso_date(request.GET.get('start_date')) or default_start_day
-    selected_end_day = _parse_iso_date(request.GET.get('end_date')) or today
-    if selected_start_day > selected_end_day:
-        selected_start_day, selected_end_day = selected_end_day, selected_start_day
-    return selected_start_day, selected_end_day, today
-
-
-def _datetime_bounds(start_day: date, end_day: date) -> tuple[datetime, datetime]:
-    """Return timezone-aware inclusive-start/exclusive-end range bounds."""
-    current_timezone = timezone.get_current_timezone()
-    start_dt = timezone.make_aware(
-        datetime.combine(start_day, time.min),
-        current_timezone,
-    )
-    end_dt = timezone.make_aware(
-        datetime.combine(end_day + timedelta(days=1), time.min),
-        current_timezone,
-    )
-    return start_dt, end_dt
-
-
-def _bool_query_value(value: bool) -> str:
-    """Return the lowercase boolean strings used by dashboard query params."""
-    return 'true' if value else 'false'
-
-
-def _metrics_query_string(
-    start_day: date,
-    end_day: date,
-    persist_mode: bool,
-    distinct_mode: bool,
-    host_sort: str,
-    trend_sort: str,
-) -> str:
-    """Build a stable query string that preserves dashboard filter state."""
-    return '?' + urlencode({
-        'start_date': start_day.isoformat(),
-        'end_date': end_day.isoformat(),
-        'persist_mode': _bool_query_value(persist_mode),
-        'distinct_mode': _bool_query_value(distinct_mode),
-        'host_sort': host_sort,
-        'trend_sort': trend_sort,
-    })
-
-
-def _dedupe_findings_queryset(queryset: QuerySet) -> QuerySet:
-    """Deduplicate repeated scan rows for the same host + service + finding.
-
-    The representative row is the highest primary key in each identity group.
-    This keeps the queryset inexpensive and deterministic for dashboard counts
-    while preserving the original model rows.
-    """
-    dedup_ids = (
-        queryset.values('host_ipv4', 'port', 'proto', 'nvt_oid', 'nvt_name')
-        .annotate(representative_id=Max('id'))
-        .values_list('representative_id', flat=True)
-    )
-    return queryset.filter(id__in=dedup_ids)
-
-
-def _clean_metric_value(value) -> str:
-    """Normalize nullable DB values for stable comparison keys."""
-    if value is None:
-        return ''
-    return str(value).strip()
-
-
-def _finding_identity(row: dict) -> tuple[str, str, str]:
-    """Return the stable per-host finding identity used across scan reports."""
-    finding_key = (
-        _clean_metric_value(row.get('nvt_oid'))
-        or _clean_metric_value(row.get('nvt_name'))
-        or _clean_metric_value(row.get('vulnerability_name'))
-        or f"uuid:{_clean_metric_value(row.get('uuid'))}"
-    )
-    return (
-        finding_key,
-        _clean_metric_value(row.get('port')).lower(),
-        _clean_metric_value(row.get('proto')).lower(),
-    )
-
-
-def _event_finding_key(event: dict) -> tuple[str, str]:
-    """Group mitigation events the same way top finding rows are grouped."""
-    return (
-        _clean_metric_value(event.get('nvt_oid')),
-        _clean_metric_value(event.get('nvt_name')),
-    )
-
-
-def _build_mitigation_events_from_rows(rows) -> list[dict]:
-    """Build mitigation events by comparing consecutive same-host scan reports.
-
-    A finding is considered mitigated when it appears in an older report and is
-    absent from the next comparable report for the same host and scanner task.
-    The task id is part of the comparison key so manual, registration and
-    periodic scans are not mixed with each other.
-    """
-    grouped_reports = defaultdict(dict)
-
-    for row in rows:
-        host_ipv4 = _clean_metric_value(row.get('host_ipv4'))
-        task_id = _clean_metric_value(row.get('task_id'))
-        report_id = _clean_metric_value(row.get('report_id'))
-        report_date = row.get('date_time')
-        if not host_ipv4 or not task_id or not report_id or report_date is None:
-            continue
-
-        identity = _finding_identity(row)
-        if not identity[0]:
-            continue
-
-        group_key = (host_ipv4, task_id)
-        report = grouped_reports[group_key].setdefault(
-            report_id,
-            {
-                'report_id': report_id,
-                'date_time': report_date,
-                'findings': {},
-            },
-        )
-        if report_date > report['date_time']:
-            report['date_time'] = report_date
-
-        report['findings'].setdefault(
-            identity,
-            {
-                'host_ipv4': host_ipv4,
-                'task_id': task_id,
-                'source_report_id': report_id,
-                'identity': identity,
-                'nvt_oid': _clean_metric_value(row.get('nvt_oid')),
-                'nvt_name': _clean_metric_value(row.get('nvt_name')),
-                'vulnerability_name': _clean_metric_value(row.get('vulnerability_name')),
-                'port': _clean_metric_value(row.get('port')),
-                'proto': _clean_metric_value(row.get('proto')),
-                'cvss_base_score': row.get('cvss_base_score'),
-            },
-        )
-
-    events = []
-    for (_host_ipv4, _task_id), report_map in grouped_reports.items():
-        reports = sorted(
-            report_map.values(),
-            key=lambda report: (report['date_time'], report['report_id']),
-        )
-        previous_findings = None
-        open_events = {}
-
-        for report in reports:
-            current_findings = report['findings']
-            if previous_findings is None:
-                previous_findings = current_findings
-                continue
-
-            previous_identities = set(previous_findings.keys())
-            current_identities = set(current_findings.keys())
-
-            for identity in current_identities - previous_identities:
-                open_event = open_events.pop(identity, None)
-                if open_event is not None:
-                    open_event['reappeared_at'] = report['date_time']
-                    open_event['reappeared_report_id'] = report['report_id']
-
-            for identity in previous_identities - current_identities:
-                event = dict(previous_findings[identity])
-                event['mitigated_at'] = report['date_time']
-                event['mitigated_report_id'] = report['report_id']
-                event['reappeared_at'] = None
-                event['reappeared_report_id'] = ''
-                events.append(event)
-                open_events[identity] = event
-
-            previous_findings = current_findings
-
-    return events
-
-
-def _build_mitigation_events(queryset: QuerySet) -> list[dict]:
-    """Query the fields needed for scan comparison and return mitigation events."""
-    rows = queryset.exclude(task_id='').exclude(report_id='').values(
-        'host_ipv4',
-        'task_id',
-        'report_id',
-        'date_time',
-        'uuid',
-        'vulnerability_name',
-        'nvt_name',
-        'nvt_oid',
-        'port',
-        'proto',
-        'cvss_base_score',
-    ).order_by('host_ipv4', 'task_id', 'date_time', 'report_id')
-    return _build_mitigation_events_from_rows(rows)
-
-
-def _event_is_currently_mitigated(event: dict, as_of_dt: datetime) -> bool:
-    """Return True if the disappearance still holds at the snapshot time."""
-    mitigated_at = event.get('mitigated_at')
-    reappeared_at = event.get('reappeared_at')
-    return (
-        mitigated_at is not None
-        and mitigated_at < as_of_dt
-        and (reappeared_at is None or reappeared_at >= as_of_dt)
-    )
-
-
-def _select_mitigation_events(
-    events: list[dict],
-    start_dt: datetime,
-    end_dt: datetime,
-    persist_mode: bool,
-) -> list[dict]:
-    """Select mitigation events for the dashboard mode and time window."""
-    if persist_mode:
-        return [event for event in events if _event_is_currently_mitigated(event, end_dt)]
-    return [
-        event for event in events
-        if event.get('mitigated_at') is not None
-        and start_dt <= event.get('mitigated_at') < end_dt
-    ]
-
-
-def _mitigation_breakdowns(events: list[dict]) -> tuple[Counter, Counter]:
-    """Return mitigation counts grouped by host and by finding."""
-    return (
-        Counter(_clean_metric_value(event.get('host_ipv4')) for event in events),
-        Counter(_event_finding_key(event) for event in events),
-    )
-
-
-def vulnerability_metrics_view(request):
-    """Build context for the custom admin metrics dashboard.
-
-    The view intentionally computes everything at request time from
-    ``Vulnerability`` rows (no materialized metrics table). This keeps the
-    data model simple and guarantees the dashboard reflects current DB state.
-    """
-    db_error = None
-    selected_start_day, selected_end_day, today = _selected_date_range(request)
-
-    persist_mode = request.GET.get('persist_mode', 'false').lower() == 'true'
-    distinct_mode = request.GET.get('distinct_mode', 'false').lower() == 'true'
-    host_sort = request.GET.get('host_sort', 'total')
-    trend_sort = request.GET.get('trend_sort', 'day')
-
-    start_dt, end_dt = _datetime_bounds(selected_start_day, selected_end_day)
-
-    selected_range_label = _format_range_label(selected_start_day, selected_end_day, today)
-    preset_ranges = [
-        {
-            'label': 'Today',
-            'start': today,
-            'end': today,
+def _virtual_admin_model(name: str, object_name: str, url: str) -> dict:
+    """Build one fake admin model entry for the Django admin sidebar."""
+    return {
+        'name': name,
+        'object_name': object_name,
+        'admin_url': url,
+        'add_url': None,
+        'view_only': True,
+        'perms': {
+            'add': False,
+            'change': False,
+            'delete': False,
+            'view': True,
         },
-        {
-            'label': 'Last 7 days',
-            'start': today - timedelta(days=6),
-            'end': today,
-        },
-        {
-            'label': 'Last month',
-            'start': today - timedelta(days=29),
-            'end': today,
-        },
-        {
-            'label': 'Last year',
-            'start': today - timedelta(days=364),
-            'end': today,
-        },
-    ]
-    total_count = 0
-    active_count = 0
-    silenced_count = 0
-    mitigated_count = 0
-    hosts_count = 0
-    reports_count = 0
-    latest_detection = None
-    critical_count = 0
-    high_count = 0
-    metric_scope_label = ''
-    severity_rows = []
-    top_hosts = []
-    top_findings = []
-    trend_rows = []
-    severity_segments = []
-    trend_total_segments = []
-    trend_active_segments = []
-    trend_chart = {
-        'width': TREND_CHART_WIDTH,
-        'height': TREND_CHART_HEIGHT,
-        'total_points': '',
-        'active_points': '',
-        'y_grid': [],
-        'x_labels': [],
     }
 
-    try:
-        # Always query the dedicated vulnerability database, not default.
-        all_vulnerabilities = Vulnerability.objects.using('vulnerability_mgmt').all()
-        range_vulnerabilities = all_vulnerabilities.filter(
-            date_time__gte=start_dt,
-            date_time__lt=end_dt,
-        )
 
-        # Scans mode: metrics are computed on selected range only.
-        # Persisted mode: metrics represent the end-of-range snapshot.
-        vulnerabilities = (
-            all_vulnerabilities.filter(date_time__lt=end_dt)
-            if persist_mode else range_vulnerabilities
-        )
-        if distinct_mode:
-            vulnerabilities = _dedupe_findings_queryset(vulnerabilities)
-            range_vulnerabilities = _dedupe_findings_queryset(range_vulnerabilities)
-
-        mitigation_events = _build_mitigation_events(
-            all_vulnerabilities.filter(date_time__lt=end_dt)
-        )
-        selected_mitigation_events = _select_mitigation_events(
-            mitigation_events,
-            start_dt,
-            end_dt,
-            persist_mode,
-        )
-        mitigated_count = len(selected_mitigation_events)
-        mitigated_by_host, mitigated_by_finding = _mitigation_breakdowns(
-            selected_mitigation_events
-        )
-
-        # Collect core counters and severity buckets in one DB round-trip.
-        # This keeps the page responsive on larger datasets.
-        aggregated = vulnerabilities.aggregate(
-            total=Count('id'),
-            active=Count('id', filter=Q(is_silenced=False)),
-            silenced=Count('id', filter=Q(is_silenced=True)),
-            hosts=Count('host_ipv4', distinct=True),
-            reports=Count(
-                'report_id',
-                distinct=True,
-                filter=~Q(report_id='') & Q(report_id__isnull=False),
-            ),
-            sev_critical=Count('id', filter=Q(cvss_base_score__gte=9.0)),
-            sev_high=Count(
-                'id',
-                filter=Q(cvss_base_score__gte=7.0, cvss_base_score__lt=9.0),
-            ),
-            sev_medium=Count(
-                'id',
-                filter=Q(cvss_base_score__gte=4.0, cvss_base_score__lt=7.0),
-            ),
-            sev_low=Count(
-                'id',
-                filter=Q(cvss_base_score__gt=0.0, cvss_base_score__lt=4.0),
-            ),
-            sev_unknown=Count(
-                'id',
-                filter=Q(cvss_base_score=0.0) | Q(cvss_base_score__isnull=True),
-            ),
-        )
-
-        # Top-level counters shown in dashboard cards.
-        total_count = aggregated.get('total') or 0
-        active_count = aggregated.get('active') or 0
-        silenced_count = aggregated.get('silenced') or 0
-        hosts_count = aggregated.get('hosts') or 0
-        reports_count = aggregated.get('reports') or 0
-        latest_detection = (
-            vulnerabilities.order_by('-date_time')
-            .values_list('date_time', flat=True)
-            .first()
-        )
-        critical_count = aggregated.get('sev_critical') or 0
-        high_count = aggregated.get('sev_high') or 0
-        metric_scope_label = (
-            f"Snapshot through {selected_end_day.strftime('%d.%m.%Y')}"
-            if persist_mode else 'Within selected range'
-        )
-
-        # Severity buckets with colors shared by both table bars and SVG chart.
-        severity_rows = [
-            {
-                'label': 'Critical (CVSS >= 9.0)',
-                'count': critical_count,
-                'color': COLOR_CRITICAL,
-            },
-            {
-                'label': 'High (CVSS 7.0 - 8.9)',
-                'count': high_count,
-                'color': COLOR_HIGH,
-            },
-            {
-                'label': 'Medium (CVSS 4.0 - 6.9)',
-                'count': aggregated.get('sev_medium') or 0,
-                'color': COLOR_MEDIUM,
-            },
-            {
-                'label': 'Low (CVSS 0.1 - 3.9)',
-                'count': aggregated.get('sev_low') or 0,
-                'color': COLOR_LOW,
-            },
-            {
-                'label': 'None/Unknown (CVSS = 0.0 or NULL)',
-                'count': aggregated.get('sev_unknown') or 0,
-                'color': COLOR_UNKNOWN,
-            },
-        ]
-        # Normalize counts to percentages for compact bar rendering in templates.
-        for row in severity_rows:
-            row['pct'] = int((row['count'] / total_count) * 100) if total_count > 0 else 0
-
-        # Build stacked-bar segments for the severity SVG chart.
-        # The SVG background rect starts at x=18; segments must match that
-        # offset. The offset is applied here in Python so the template can
-        # render {{ seg.x }} directly without |add:'18', which would truncate
-        # float values to int and cause sub-pixel misalignment.
-        segment_x = float(SEVERITY_CHART_OFFSET)
-        for row in severity_rows:
-            seg_width = (
-                (row['count'] / total_count) * SEVERITY_CHART_WIDTH
-                if total_count > 0 else 0
-            )
-            severity_segments.append({
-                'x': round(segment_x, 2),
-                'width': round(seg_width, 2),
-                'color': row['color'],
-            })
-            # Keep a running x-offset so segments are contiguous.
-            segment_x += seg_width
-
-        host_order_by = ['-total', 'host_ipv4']
-        if host_sort == 'max_cvss':
-            host_order_by = ['-max_cvss', '-total', 'host_ipv4']
-        elif host_sort == 'active':
-            host_order_by = ['-active', '-total', 'host_ipv4']
-        elif host_sort == 'silenced':
-            host_order_by = ['-silenced', '-total', 'host_ipv4']
-        elif host_sort == 'mitigated':
-            host_order_by = ['-total', 'host_ipv4']
-        elif host_sort == 'host':
-            host_order_by = ['host_ipv4']
-        else:
-            host_sort = 'total'
-
-        host_rows = (
-            vulnerabilities.values('host_ipv4')
-            .annotate(
-                total=Count('id'),
-                active=Count('id', filter=Q(is_silenced=False)),
-                silenced=Count('id', filter=Q(is_silenced=True)),
-                max_cvss=Max('cvss_base_score'),
-                active_critical=Count(
-                    'id',
-                    filter=Q(is_silenced=False, cvss_base_score__gte=9.0),
-                ),
-                active_high=Count(
-                    'id',
-                    filter=Q(is_silenced=False, cvss_base_score__gte=7.0, cvss_base_score__lt=9.0),
-                ),
-                distinct_findings=Count('nvt_oid', distinct=True),
-                reports=Count(
-                    'report_id',
-                    distinct=True,
-                    filter=~Q(report_id='') & Q(report_id__isnull=False),
-                ),
-            )
-            .order_by(*host_order_by)
-        )
-        top_hosts = list(host_rows if host_sort == 'mitigated' else host_rows[:10])
-        for host in top_hosts:
-            host['mitigated'] = mitigated_by_host.get(
-                _clean_metric_value(host.get('host_ipv4')),
-                0,
-            )
-        if host_sort == 'mitigated':
-            # Stable two-step sort: deterministic host ordering first, then by
-            # mitigated/total counts descending for reviewer-friendly ranking.
-            top_hosts.sort(
-                key=lambda host: (
-                    _clean_metric_value(host.get('host_ipv4')),
-                )
-            )
-            top_hosts.sort(
-                key=lambda host: (
-                    host.get('mitigated', 0),
-                    host.get('total', 0),
-                ),
-                reverse=True,
-            )
-            top_hosts = top_hosts[:10]
-        # Host load bars are relative to the most affected host, not global total.
-        host_peak = max((host['total'] for host in top_hosts), default=0)
-        for host in top_hosts:
-            host['pct'] = int((host['total'] / host_peak) * 100) if host_peak > 0 else 0
-            host['bar_color'] = _cvss_bar_color(host.get('max_cvss'))
-            host['hover_info'] = (
-                f"Active critical: {host.get('active_critical', 0)}, "
-                f"active high: {host.get('active_high', 0)}, "
-                f"silenced: {host.get('silenced', 0)}, "
-                f"mitigated: {host.get('mitigated', 0)}, "
-                f"unique NVT OIDs: {host.get('distinct_findings', 0)}, "
-                f"reports: {host.get('reports', 0)}"
-            )
-
-        # Rank most frequent findings (NVT names).
-        top_findings = list(
-            vulnerabilities.exclude(nvt_name='')
-            .values('nvt_name', 'nvt_oid')
-            .annotate(
-                total=Count('id'),
-                active=Count('id', filter=Q(is_silenced=False)),
-                silenced=Count('id', filter=Q(is_silenced=True)),
-                host_count=Count('host_ipv4', distinct=True),
-                report_count=Count(
-                    'report_id',
-                    distinct=True,
-                    filter=~Q(report_id='') & Q(report_id__isnull=False),
-                ),
-                sample_uuid=Max('uuid'),
-                max_cvss=Max('cvss_base_score'),
-            )
-            .order_by('-total', 'nvt_name')[:10]
-        )
-        # Finding bars are relative to the most common finding.
-        finding_peak = max((finding['total'] for finding in top_findings), default=0)
-        for finding in top_findings:
-            finding['mitigated'] = mitigated_by_finding.get(
-                (
-                    _clean_metric_value(finding.get('nvt_oid')),
-                    _clean_metric_value(finding.get('nvt_name')),
-                ),
-                0,
-            )
-            finding['pct'] = (
-                int((finding['total'] / finding_peak) * 100)
-                if finding_peak > 0 else 0
-            )
-            finding['bar_color'] = _cvss_bar_color(finding.get('max_cvss'))
-            finding['hover_info'] = (
-                f"Distinct hosts: {finding.get('host_count', 0)}, "
-                f"reports: {finding.get('report_count', 0)}, "
-                f"active: {finding.get('active', 0)}, "
-                f"silenced: {finding.get('silenced', 0)}, "
-                f"mitigated: {finding.get('mitigated', 0)}, "
-                f"max CVSS: {finding.get('max_cvss', '-') or '-'}"
-            )
-
-        # Build the selected time window so the trend chart matches the picker.
-        start_day = selected_start_day
-        end_day = selected_end_day
-
-        if persist_mode:
-            # Persisted mode: for each day, count all vulnerabilities that existed on that day.
-            # A vulnerability "existed" on a day if it was detected before end-of-day.
-            trend_query = []
-            for day_offset in range((end_day - start_day).days + 1):
-                current_day = start_day + timedelta(days=day_offset)
-                day_end_dt = timezone.make_aware(
-                    datetime.combine(current_day + timedelta(days=1), time.min),
-                    timezone.get_current_timezone(),
-                )
-                day_vulns = all_vulnerabilities.filter(
-                    date_time__lt=day_end_dt,
-                )
-                if distinct_mode:
-                    day_vulns = _dedupe_findings_queryset(day_vulns)
-                day_agg = day_vulns.aggregate(
-                    total=Count('id'),
-                    active=Count('id', filter=Q(is_silenced=False)),
-                    silenced=Count('id', filter=Q(is_silenced=True)),
-                    critical=Count('id', filter=Q(cvss_base_score__gte=9.0)),
-                    high=Count(
-                        'id',
-                        filter=Q(cvss_base_score__gte=7.0, cvss_base_score__lt=9.0),
-                    ),
-                )
-                trend_query.append({
-                    'day': current_day,
-                    'total': day_agg['total'] or 0,
-                    'active': day_agg['active'] or 0,
-                    'silenced': day_agg['silenced'] or 0,
-                    'mitigated': len(_select_mitigation_events(
-                        mitigation_events,
-                        start_dt,
-                        day_end_dt,
-                        persist_mode,
-                    )),
-                    'critical': day_agg['critical'] or 0,
-                    'high': day_agg['high'] or 0,
-                })
-        else:
-            # Scan mode: only findings detected within each calendar day.
-            # Use the same explicit day boundaries as persisted mode to avoid
-            # timezone/day-bucket drift between both modes.
-            trend_query = []
-            for day_offset in range((end_day - start_day).days + 1):
-                current_day = start_day + timedelta(days=day_offset)
-                day_start_dt = timezone.make_aware(
-                    datetime.combine(current_day, time.min),
-                    timezone.get_current_timezone(),
-                )
-                day_end_dt = timezone.make_aware(
-                    datetime.combine(current_day + timedelta(days=1), time.min),
-                    timezone.get_current_timezone(),
-                )
-                day_vulns = range_vulnerabilities.filter(
-                    date_time__gte=day_start_dt,
-                    date_time__lt=day_end_dt,
-                )
-                day_agg = day_vulns.aggregate(
-                    total=Count('id'),
-                    active=Count('id', filter=Q(is_silenced=False)),
-                    silenced=Count('id', filter=Q(is_silenced=True)),
-                    critical=Count('id', filter=Q(cvss_base_score__gte=9.0)),
-                    high=Count(
-                        'id',
-                        filter=Q(cvss_base_score__gte=7.0, cvss_base_score__lt=9.0),
-                    ),
-                )
-                trend_query.append({
-                    'day': current_day,
-                    'total': day_agg['total'] or 0,
-                    'active': day_agg['active'] or 0,
-                    'silenced': day_agg['silenced'] or 0,
-                    'mitigated': len(_select_mitigation_events(
-                        mitigation_events,
-                        day_start_dt,
-                        day_end_dt,
-                        persist_mode,
-                    )),
-                    'critical': day_agg['critical'] or 0,
-                    'high': day_agg['high'] or 0,
-                })
-
-        trend_map = {row['day']: row for row in trend_query}
-
-        # Fill missing days with zeros so x-axis spacing remains stable.
-        trend_peak = 0
-        days_span = (end_day - start_day).days + 1
-        for day_offset in range(days_span):
-            day = start_day + timedelta(days=day_offset)
-            row = trend_map.get(day, {
-                'total': 0,
-                'active': 0,
-                'silenced': 0,
-                'mitigated': 0,
-                'critical': 0,
-                'high': 0,
-            })
-            severity_color = _severity_tint(row['critical'], row['high'], row['total'])
-            trend_rows.append({
-                'day': day,
-                'total': row['total'],
-                'active': row['active'],
-                'silenced': row['silenced'],
-                'mitigated': row['mitigated'],
-                'critical': row['critical'],
-                'high': row['high'],
-                'severity_color': severity_color,
-            })
-            trend_peak = max(trend_peak, row['total'])
-
-        for row in trend_rows:
-            row['total_pct'] = (
-                int((row['total'] / trend_peak) * 100)
-                if trend_peak > 0 else 0
-            )
-            row['active_pct'] = (
-                int((row['active'] / trend_peak) * 100)
-                if trend_peak > 0 else 0
-            )
-
-        for row in trend_rows:
-            # Hover payloads use per-day host/report aggregates so tooltip
-            # details stay aligned with whichever mode/range is selected.
-            day_start_dt = timezone.make_aware(
-                datetime.combine(row['day'], time.min),
-                timezone.get_current_timezone(),
-            )
-            day_end_dt = timezone.make_aware(
-                datetime.combine(row['day'] + timedelta(days=1), time.min),
-                timezone.get_current_timezone(),
-            )
-            if persist_mode:
-                day_vulns = all_vulnerabilities.filter(
-                    date_time__lt=day_end_dt,
-                )
-                if distinct_mode:
-                    day_vulns = _dedupe_findings_queryset(day_vulns)
-            else:
-                day_vulns = range_vulnerabilities.filter(
-                    date_time__gte=day_start_dt,
-                    date_time__lt=day_end_dt,
-                )
-
-            day_stats = day_vulns.aggregate(
-                distinct_hosts=Count('host_ipv4', distinct=True),
-                reports=Count(
-                    'report_id',
-                    distinct=True,
-                    filter=~Q(report_id='') & Q(report_id__isnull=False),
-                ),
-            )
-            top_hosts_for_day = list(
-                day_vulns.values('host_ipv4')
-                .annotate(
-                    total=Count('id'),
-                    critical=Count('id', filter=Q(cvss_base_score__gte=9.0)),
-                    high=Count(
-                        'id',
-                        filter=Q(cvss_base_score__gte=7.0, cvss_base_score__lt=9.0),
-                    ),
-                )
-                .order_by('-total', 'host_ipv4')[:3]
-            )
-            if top_hosts_for_day:
-                top_hosts_text = ', '.join(
-                    f"{host['host_ipv4']} ({host['total']}, C{host['critical']}/H{host['high']})"
-                    for host in top_hosts_for_day
-                )
-            else:
-                top_hosts_text = 'No host breakdown available'
-
-            row['hover_info'] = (
-                f"Distinct hosts: {day_stats.get('distinct_hosts', 0)}, "
-                f"reports: {day_stats.get('reports', 0)}, "
-                f"silenced: {row.get('silenced', 0)}, "
-                f"mitigated: {row.get('mitigated', 0)}, "
-                f"top hosts: {top_hosts_text}"
-            )
-
-        trend_table_rows = list(trend_rows)
-        if trend_sort == 'day':
-            trend_table_rows.sort(
-                key=lambda row: row['day'],
-                reverse=True,
-            )
-        elif trend_sort == 'critical':
-            trend_table_rows.sort(
-                key=lambda row: (row['critical'], row['total'], row['day']),
-                reverse=True,
-            )
-        elif trend_sort == 'high':
-            trend_table_rows.sort(
-                key=lambda row: (row['high'], row['critical'], row['total'], row['day']),
-                reverse=True,
-            )
-        elif trend_sort == 'active':
-            trend_table_rows.sort(
-                key=lambda row: (row['active'], row['critical'], row['total'], row['day']),
-                reverse=True,
-            )
-        elif trend_sort == 'silenced':
-            trend_table_rows.sort(
-                key=lambda row: (row['silenced'], row['critical'], row['total'], row['day']),
-                reverse=True,
-            )
-        elif trend_sort == 'mitigated':
-            trend_table_rows.sort(
-                key=lambda row: (row['mitigated'], row['critical'], row['total'], row['day']),
-                reverse=True,
-            )
-        elif trend_sort == 'total':
-            trend_table_rows.sort(
-                key=lambda row: (row['total'], row['critical'], row['day']),
-                reverse=True,
-            )
-        else:
-            trend_sort = 'day'
-            trend_table_rows.sort(
-                key=lambda row: row['day'],
-                reverse=True,
-            )
-
-        # Convert trend rows into chart coordinates for inline SVG rendering.
-        # Plot dimensions represent the drawable area inside padded axes.
-        plot_w = TREND_CHART_WIDTH - TREND_CHART_LEFT_PAD - TREND_CHART_RIGHT_PAD
-        plot_h = TREND_CHART_HEIGHT - TREND_CHART_TOP_PAD - TREND_CHART_BOTTOM_PAD
-        point_count = len(trend_rows)
-        x_step = (plot_w / (point_count - 1)) if point_count > 1 else 0
-
-        total_pts = []
-        active_pts = []
-        for idx, row in enumerate(trend_rows):
-            x = TREND_CHART_LEFT_PAD + (idx * x_step)
-            y_total = TREND_CHART_TOP_PAD + plot_h
-            y_active = TREND_CHART_TOP_PAD + plot_h
-            if trend_peak > 0:
-                y_total = TREND_CHART_TOP_PAD + plot_h - ((row['total'] / trend_peak) * plot_h)
-                y_active = TREND_CHART_TOP_PAD + plot_h - ((row['active'] / trend_peak) * plot_h)
-
-            row['chart_x'] = round(x, 2)
-            row['chart_total_y'] = round(y_total, 2)
-            row['chart_active_y'] = round(y_active, 2)
-            row['chart_total_h'] = round((TREND_CHART_TOP_PAD + plot_h) - y_total, 2)
-            row['chart_active_h'] = round((TREND_CHART_TOP_PAD + plot_h) - y_active, 2)
-            total_pts.append(f"{row['chart_x']},{row['chart_total_y']}")
-            active_pts.append(f"{row['chart_x']},{row['chart_active_y']}")
-
-        for idx in range(len(trend_rows) - 1):
-            current_row = trend_rows[idx]
-            next_row = trend_rows[idx + 1]
-            segment_color = next_row['severity_color']
-            trend_total_segments.append({
-                'x1': current_row['chart_x'],
-                'y1': current_row['chart_total_y'],
-                'x2': next_row['chart_x'],
-                'y2': next_row['chart_total_y'],
-                'color': segment_color,
-            })
-            trend_active_segments.append({
-                'x1': current_row['chart_x'],
-                'y1': current_row['chart_active_y'],
-                'x2': next_row['chart_x'],
-                'y2': next_row['chart_active_y'],
-                'color': segment_color,
-            })
-
-        y_grid = []
-        for i in range(5):
-            ratio = i / 4
-            y = TREND_CHART_TOP_PAD + plot_h - (ratio * plot_h)
-            y_grid.append({
-                'y': round(y, 2),
-                'value': int(round(trend_peak * ratio)),
-            })
-
-        x_labels = []
-        for idx, row in enumerate(trend_rows):
-            # Show a subset of labels to keep the x-axis readable.
-            show_every = max(1, round(point_count / 4))
-            if idx == 0 or idx == point_count - 1 or idx % show_every == 0:
-                x_labels.append({
-                    'x': row['chart_x'],
-                    'label': row['day'].strftime('%m-%d'),
-                })
-
-        trend_chart = {
-            'width': TREND_CHART_WIDTH,
-            'height': TREND_CHART_HEIGHT,
-            'left_pad': TREND_CHART_LEFT_PAD,
-            'right_pad': TREND_CHART_RIGHT_PAD,
-            'top_pad': TREND_CHART_TOP_PAD,
-            'bottom_pad': TREND_CHART_BOTTOM_PAD,
-            'plot_w': plot_w,
-            'plot_h': plot_h,
-            'total_points': ' '.join(total_pts),
-            'active_points': ' '.join(active_pts),
-            'total_segments': trend_total_segments,
-            'active_segments': trend_active_segments,
-            'y_grid': y_grid,
-            'x_labels': x_labels,
-        }
-
-    except (ProgrammingError, OperationalError):
-        # Keep the admin page readable when migration/setup was not run yet.
-        db_error = (
-            'Vulnerability table not found. '
-            'Run: python manage.py migrate --database=vulnerability_mgmt'
-        )
-    except Exception:
-        # Keep the admin page usable even if a non-DB assumption breaks.
-        logger.exception('Unexpected error while building vulnerability metrics.')
-        db_error = 'Could not build vulnerability metrics due to an unexpected error.'
-
-    toggle_persist_query = _metrics_query_string(
-        selected_start_day,
-        selected_end_day,
-        not persist_mode,
-        distinct_mode,
-        host_sort,
-        trend_sort,
-    )
-    toggle_distinct_query = _metrics_query_string(
-        selected_start_day,
-        selected_end_day,
-        persist_mode,
-        not distinct_mode,
-        host_sort,
-        trend_sort,
-    )
-    host_detail_query = _metrics_query_string(
-        selected_start_day,
-        selected_end_day,
-        persist_mode,
-        distinct_mode,
-        host_sort,
-        trend_sort,
-    )
-    for preset in preset_ranges:
-        # Precompute links so templates can render compact, readable anchors.
-        preset['query'] = _metrics_query_string(
-            preset['start'],
-            preset['end'],
-            persist_mode,
-            distinct_mode,
-            host_sort,
-            trend_sort,
-        )
-    host_sort_queries = {
-        sort_key: _metrics_query_string(
-            selected_start_day,
-            selected_end_day,
-            persist_mode,
-            distinct_mode,
-            sort_key,
-            trend_sort,
-        )
-        for sort_key in ('total', 'active', 'silenced', 'mitigated', 'max_cvss')
-    }
-    trend_sort_queries = {
-        sort_key: _metrics_query_string(
-            selected_start_day,
-            selected_end_day,
-            persist_mode,
-            distinct_mode,
-            host_sort,
-            sort_key,
-        )
-        for sort_key in ('day', 'total', 'active', 'silenced', 'mitigated', 'critical', 'high')
-    }
-
-    context = {
-        **admin.site.each_context(request),
-        'title': 'Vulnerability Metrics',
-        'db_error': db_error,
-        'active_count': active_count,
-        'silenced_count': silenced_count,
-        'mitigated_count': mitigated_count,
-        'hosts_count': hosts_count,
-        'critical_count': critical_count,
-        'high_count': high_count,
-        'reports_count': reports_count,
-        'latest_detection': latest_detection,
-        'total_count': total_count,
-        'metric_scope_label': metric_scope_label,
-        'severity_rows': severity_rows,
-        'top_hosts': top_hosts,
-        'top_findings': top_findings,
-        'trend_rows': trend_rows,
-        'trend_table_rows': trend_table_rows if 'trend_table_rows' in locals() else trend_rows,
-        'severity_segments': severity_segments,
-        'trend_chart': trend_chart,
-        'selected_start_day': selected_start_day,
-        'selected_end_day': selected_end_day,
-        'selected_start_value': selected_start_day.isoformat(),
-        'selected_end_value': selected_end_day.isoformat(),
-        'selected_range_label': selected_range_label,
-        'preset_ranges': preset_ranges,
-        'persist_mode': persist_mode,
-        'distinct_mode': distinct_mode,
-        'host_sort': host_sort,
-        'trend_sort': trend_sort,
-        'toggle_persist_query': toggle_persist_query,
-        'toggle_distinct_query': toggle_distinct_query,
-        'host_detail_query': host_detail_query,
-        'host_sort_queries': host_sort_queries,
-        'trend_sort_queries': trend_sort_queries,
-    }
-    return TemplateResponse(request, 'admin/vulnerability_metrics.html', context)
-
-
-def vulnerability_host_detail_view(request, host_ipv4: str):
-    """Detailed host view linked from top affected hosts table."""
-    db_error = None
-    selected_start_day, selected_end_day, today = _selected_date_range(request)
-
-    persist_mode = request.GET.get('persist_mode', 'false').lower() == 'true'
-    distinct_mode = request.GET.get('distinct_mode', 'false').lower() == 'true'
-    host_sort = request.GET.get('host_sort', 'total')
-    trend_sort = request.GET.get('trend_sort', 'day')
-
-    start_dt, end_dt = _datetime_bounds(selected_start_day, selected_end_day)
-
-    selected_range_label = _format_range_label(selected_start_day, selected_end_day, today)
-    host_summary = {
-        'total': 0,
-        'active': 0,
-        'silenced': 0,
-        'mitigated': 0,
-        'max_cvss': None,
-        'critical': 0,
-        'high': 0,
-    }
-    host_top_findings = []
-    host_recent_vulns = []
-
-    try:
-        all_host_vulns = (
-            Vulnerability.objects.using('vulnerability_mgmt')
-            .filter(host_ipv4=host_ipv4)
-        )
-        range_host_vulns = all_host_vulns.filter(date_time__gte=start_dt, date_time__lt=end_dt)
-        host_vulns = (
-            all_host_vulns.filter(date_time__lt=end_dt)
-            if persist_mode else range_host_vulns
-        )
-        if distinct_mode:
-            host_vulns = _dedupe_findings_queryset(host_vulns)
-
-        host_mitigation_events = _build_mitigation_events(
-            all_host_vulns.filter(date_time__lt=end_dt)
-        )
-        selected_host_mitigations = _select_mitigation_events(
-            host_mitigation_events,
-            start_dt,
-            end_dt,
-            persist_mode,
-        )
-        _mitigated_by_host, host_mitigated_by_finding = _mitigation_breakdowns(
-            selected_host_mitigations
-        )
-
-        aggregated = host_vulns.aggregate(
-            total=Count('id'),
-            active=Count('id', filter=Q(is_silenced=False)),
-            silenced=Count('id', filter=Q(is_silenced=True)),
-            max_cvss=Max('cvss_base_score'),
-            critical=Count('id', filter=Q(cvss_base_score__gte=9.0)),
-            high=Count('id', filter=Q(cvss_base_score__gte=7.0, cvss_base_score__lt=9.0)),
-        )
-        host_summary = {
-            'total': aggregated.get('total') or 0,
-            'active': aggregated.get('active') or 0,
-            'silenced': aggregated.get('silenced') or 0,
-            'mitigated': len(selected_host_mitigations),
-            'max_cvss': aggregated.get('max_cvss'),
-            'critical': aggregated.get('critical') or 0,
-            'high': aggregated.get('high') or 0,
-        }
-
-        host_top_findings = list(
-            host_vulns.exclude(nvt_name='')
-            .values('nvt_name', 'nvt_oid')
-            .annotate(
-                total=Count('id'),
-                active=Count('id', filter=Q(is_silenced=False)),
-                silenced=Count('id', filter=Q(is_silenced=True)),
-                max_cvss=Max('cvss_base_score'),
-            )
-            .order_by('-total', 'nvt_name')[:25]
-        )
-        for finding in host_top_findings:
-            finding['mitigated'] = host_mitigated_by_finding.get(
-                (
-                    _clean_metric_value(finding.get('nvt_oid')),
-                    _clean_metric_value(finding.get('nvt_name')),
-                ),
-                0,
-            )
-
-        host_recent_vulns = list(
-            host_vulns.order_by('-date_time').values(
-                'date_time',
-                'vulnerability_name',
-                'nvt_name',
-                'cvss_base_score',
-                'is_silenced',
-                'port',
-                'proto',
-                'report_id',
-                'uuid',
-            )[:80]
-        )
-
-    except (ProgrammingError, OperationalError):
-        db_error = (
-            'Vulnerability table not found. '
-            'Run: python manage.py migrate --database=vulnerability_mgmt'
-        )
-    except Exception:
-        logger.exception('Unexpected error while building host details for %s.', host_ipv4)
-        db_error = 'Could not build host details due to an unexpected error.'
-
-    back_query = _metrics_query_string(
-        selected_start_day,
-        selected_end_day,
-        persist_mode,
-        distinct_mode,
-        host_sort,
-        trend_sort,
-    )
-
-    context = {
-        **admin.site.each_context(request),
-        'title': f'Host Details: {host_ipv4}',
-        'db_error': db_error,
-        'host_ipv4': host_ipv4,
-        'host_summary': host_summary,
-        'host_top_findings': host_top_findings,
-        'host_recent_vulns': host_recent_vulns,
-        'selected_range_label': selected_range_label,
-        'back_query': back_query,
-    }
-    return TemplateResponse(request, 'admin/vulnerability_host_detail.html', context)
-
-
+# Django imports admin.py during startup and autoreload. The guard prevents
+# wrapping admin.site.get_urls multiple times in the same Python process.
 if not hasattr(admin.site, '_vulnerability_metrics_urls_patched'):
-    # Guard against duplicate monkey-patching during Django reload/import cycles.
+    # Keep Django's original URL builder so default admin pages still work.
     original_get_urls = admin.site.get_urls
 
     def get_urls():
-        """Register dashboard URLs while preserving all default admin routes."""
-        # Extend default admin URLs, do not replace them.
+        # Custom admin pages live under /admin/vulnerability-metrics/.
         custom_urls = [
             path(
                 'vulnerability-metrics/',
-                admin.site.admin_view(vulnerability_metrics_view),
-                name='vulnerability_metrics'
+                # admin_view applies normal admin permission and CSRF handling.
+                admin.site.admin_view(admin_views.overview),
+                name='vulnerability_metrics',
             ),
             path(
-                'vulnerability-metrics/host/<str:host_ipv4>/',
-                admin.site.admin_view(vulnerability_host_detail_view),
-                name='vulnerability_host_detail'
+                'vulnerability-metrics/latest-reports/',
+                admin.site.admin_view(admin_views.latest_reports),
+                name='vulnerability_latest_reports',
             ),
         ]
+        # Put custom URLs first so they are matched before generic admin routes.
         return custom_urls + original_get_urls()
 
     admin.site.get_urls = get_urls
     admin.site._vulnerability_metrics_urls_patched = True
 
 
+# Same reload guard for the admin sidebar/app-list patch.
 if not hasattr(admin.site, '_vulnerability_metrics_applist_patched'):
-    # Same guard for app-list patching to avoid duplicate sidebar entries.
+    # Keep Django's original app-list builder and append our virtual section.
     original_get_app_list = admin.site.get_app_list
 
     def get_app_list(request, app_label=None):
-        """Append a virtual app section so metrics appear in admin navigation."""
-        # Append a virtual app section so the page appears in left navigation.
         app_list = original_get_app_list(request, app_label=app_label)
-        if app_label and app_label != 'vulnerability_metrics':
+        # If Django asks for a different app, do not inject this virtual app.
+        if app_label and app_label != APP_LABEL:
             return app_list
 
+        # Reverse URL names here so sidebar links follow Django's admin namespace.
         metrics_url = reverse('admin:vulnerability_metrics')
+        latest_reports_url = reverse('admin:vulnerability_latest_reports')
+        # This is not a real Django app/model. It only makes the custom pages
+        # appear in the admin navigation next to normal model admin entries.
         metrics_app = {
             'name': 'Vulnerability Metrics',
-            'app_label': 'vulnerability_metrics',
+            'app_label': APP_LABEL,
             'app_url': metrics_url,
             'has_module_perms': True,
             'models': [
-                {
-                    'name': 'Overview',
-                    'object_name': 'VulnerabilityMetrics',
-                    'admin_url': metrics_url,
-                    'add_url': None,
-                    'view_only': True,
-                    'perms': {
-                        'add': False,
-                        'change': False,
-                        'delete': False,
-                        'view': True,
-                    },
-                }
+                _virtual_admin_model(
+                    name='Overview',
+                    object_name='VulnerabilityMetrics',
+                    url=metrics_url,
+                ),
+                _virtual_admin_model(
+                    name='Latest reports',
+                    object_name='VulnerabilityLatestReports',
+                    url=latest_reports_url,
+                ),
             ],
         }
 
-        # Idempotency check: never append the same virtual app twice.
-        if not any(app.get('app_label') == 'vulnerability_metrics' for app in app_list):
+        # Avoid duplicate sidebar sections if another import/reload already added it.
+        if not any(app.get('app_label') == APP_LABEL for app in app_list):
             app_list.append(metrics_app)
         return app_list
