+++ project/application/vulnerability_mgmt/management/__init__.py
@@ -1,0 +1,1 @@
+"""Management package for vulnerability_mgmt app."""
