+++ project/application/vulnerability_mgmt/management/commands/vuln_metrics_report.py
@@ -1,0 +1,100 @@
+"""Print a compact vulnerability metrics summary to the terminal.
+
+Usage examples:
+    python manage.py vuln_metrics_report
+    python manage.py vuln_metrics_report --report-id <report_id>
+    python manage.py vuln_metrics_report --database-alias vulnerability_mgmt
+"""
+
+from django.core.management.base import BaseCommand
+from django.db.models import Avg, Count, Max
+
+from vulnerability_mgmt.models import Vulnerability
+
+
+class Command(BaseCommand):
+    help = 'Print vulnerability metrics summary from the vulnerability_mgmt DB.'
+
+    def add_arguments(self, parser):
+        parser.add_argument(
+            '--report-id',
+            default='',
+            help='Optional report_id filter for a single scan report.',
+        )
+        parser.add_argument(
+            '--database-alias',
+            default='vulnerability_mgmt',
+            help='Database alias to read from (default: vulnerability_mgmt).',
+        )
+        parser.add_argument(
+            '--limit-hosts',
+            type=int,
+            default=10,
+            help='How many top hosts to show (default: 10).',
+        )
+
+    def handle(self, *args, **options):
+        report_id = options['report_id']
+        alias = options['database_alias']
+        limit_hosts = max(1, int(options['limit_hosts']))
+
+        qs = Vulnerability.objects.using(alias).all()
+        if report_id:
+            qs = qs.filter(report_id=report_id)
+
+        total = qs.count()
+        if total == 0:
+            suffix = f' for report_id={report_id}' if report_id else ''
+            self.stdout.write(self.style.WARNING(
+                f'No vulnerability records found in alias "{alias}"{suffix}.'
+            ))
+            return
+
+        summary = qs.aggregate(
+            unique_hosts=Count('host_ipv4', distinct=True),
+            unique_reports=Count('report_id', distinct=True),
+            max_cvss=Max('cvss_base_score'),
+            avg_cvss=Avg('cvss_base_score'),
+            latest_detection=Max('date_time'),
+        )
+
+        silenced = qs.filter(is_silenced=True).count()
+        unsilenced = total - silenced
+
+        # Severity buckets based on CVSS base score.
+        critical = qs.filter(cvss_base_score__gte=9.0).count()
+        high = qs.filter(cvss_base_score__gte=7.0, cvss_base_score__lt=9.0).count()
+        medium = qs.filter(cvss_base_score__gte=4.0, cvss_base_score__lt=7.0).count()
+        low = qs.filter(cvss_base_score__gte=0.1, cvss_base_score__lt=4.0).count()
+
+        top_hosts = list(
+            qs.values('host_ipv4')
+            .annotate(findings=Count('id'))
+            .order_by('-findings', 'host_ipv4')[:limit_hosts]
+        )
+
+        self.stdout.write(self.style.SUCCESS('Vulnerability Metrics Report'))
+        self.stdout.write(f'- DB alias: {alias}')
+        self.stdout.write(f'- Total findings: {total}')
+        self.stdout.write(f'- Unsilenced findings: {unsilenced}')
+        self.stdout.write(f'- Silenced findings: {silenced}')
+        self.stdout.write(f'- Unique hosts: {summary["unique_hosts"]}')
+        self.stdout.write(f'- Unique reports: {summary["unique_reports"]}')
+        self.stdout.write(f'- Max CVSS: {summary["max_cvss"]}')
+        self.stdout.write(f'- Avg CVSS: {summary["avg_cvss"]:.2f}')
+        self.stdout.write(f'- Latest detection: {summary["latest_detection"]}')
+
+        self.stdout.write('')
+        self.stdout.write('Severity buckets (CVSS):')
+        self.stdout.write(f'- Critical (>= 9.0): {critical}')
+        self.stdout.write(f'- High (7.0 - 8.9): {high}')
+        self.stdout.write(f'- Medium (4.0 - 6.9): {medium}')
+        self.stdout.write(f'- Low (0.1 - 3.9): {low}')
+
+        if top_hosts:
+            self.stdout.write('')
+            self.stdout.write(f'Top {limit_hosts} hosts by finding count:')
+            for row in top_hosts:
+                self.stdout.write(
+                    f'- {row["host_ipv4"]}: {row["findings"]} findings'
+                )
