+++ project/vulnerability_reports_inspection.md
@@ -1,0 +1,48 @@
+# Distinct Vulnerability Reports
+
+Total report/host/task groups: 32
+Total mitigation events: 2
+
+## Reports
+
+| Report ID | Host | Task ID | First row | Last row | Findings | Mitigations caused |
+|---|---|---|---|---|---:|---:|
+| 2026-03-30T22:46:21Z | 131.173.61.9 | a02609a1-120a-4f9b-b7ee-d420aa20136f | 2026-03-30 22:42:59+00:00 | 2026-03-30 22:46:15+00:00 | 3 | 0 |
+| 2026-03-31T06:26:33Z | 131.173.22.98 | 6e809242-d12c-4aad-ab47-c4f48854383a | 2026-03-31 06:23:43+00:00 | 2026-03-31 06:26:24+00:00 | 12 | 0 |
+| 2026-04-06T22:46:28Z | 131.173.61.9 | a02609a1-120a-4f9b-b7ee-d420aa20136f | 2026-04-06 22:43:14+00:00 | 2026-04-06 22:46:25+00:00 | 3 | 0 |
+| 2026-04-07T06:27:51Z | 131.173.22.98 | 6e809242-d12c-4aad-ab47-c4f48854383a | 2026-04-07 06:23:30+00:00 | 2026-04-07 06:27:41+00:00 | 12 | 0 |
+| 2026-04-20T22:46:12Z | 131.173.61.9 | a02609a1-120a-4f9b-b7ee-d420aa20136f | 2026-04-20 22:43:33+00:00 | 2026-04-20 22:45:59+00:00 | 3 | 0 |
+| 2026-04-21T06:28:12Z | 131.173.22.98 | 6e809242-d12c-4aad-ab47-c4f48854383a | 2026-04-21 06:23:45+00:00 | 2026-04-21 06:28:01+00:00 | 12 | 0 |
+| 2026-04-22T20:25:55Z | 131.173.32.4 | 3587a126-c4dc-4312-a907-d7cd933c9db5 | 2026-04-22 20:22:36+00:00 | 2026-04-22 20:25:42+00:00 | 15 | 0 |
+| 2026-04-22T22:32:01Z | 131.173.24.54 | 730a4bed-1fa4-4143-98c0-bcff78c63013 | 2026-04-22 22:14:50+00:00 | 2026-04-22 22:31:59+00:00 | 9 | 0 |
+| 2026-04-27T22:46:59Z | 131.173.61.9 | a02609a1-120a-4f9b-b7ee-d420aa20136f | 2026-04-27 22:43:37+00:00 | 2026-04-27 22:46:52+00:00 | 3 | 0 |
+| 2026-04-28T06:27:08Z | 131.173.22.98 | 6e809242-d12c-4aad-ab47-c4f48854383a | 2026-04-28 06:24:09+00:00 | 2026-04-28 06:26:56+00:00 | 12 | 0 |
+| 2026-04-29T20:27:03Z | 131.173.32.4 | 3587a126-c4dc-4312-a907-d7cd933c9db5 | 2026-04-29 20:22:59+00:00 | 2026-04-29 20:26:55+00:00 | 15 | 0 |
+| 2026-04-29T22:22:42Z | 131.173.24.54 | 730a4bed-1fa4-4143-98c0-bcff78c63013 | 2026-04-29 22:14:00+00:00 | 2026-04-29 22:22:40+00:00 | 9 | 0 |
+| 2026-05-04T22:49:49Z | 131.173.61.9 | a02609a1-120a-4f9b-b7ee-d420aa20136f | 2026-05-04 22:45:25+00:00 | 2026-05-04 22:49:42+00:00 | 3 | 0 |
+| 2026-05-05T06:27:18Z | 131.173.22.98 | 6e809242-d12c-4aad-ab47-c4f48854383a | 2026-05-05 06:23:45+00:00 | 2026-05-05 06:27:06+00:00 | 12 | 0 |
+| 2026-05-06T20:25:35Z | 131.173.32.4 | 3587a126-c4dc-4312-a907-d7cd933c9db5 | 2026-05-06 20:22:40+00:00 | 2026-05-06 20:25:21+00:00 | 15 | 0 |
+| 2026-05-06T22:18:46Z | 131.173.24.54 | 730a4bed-1fa4-4143-98c0-bcff78c63013 | 2026-05-06 22:13:58+00:00 | 2026-05-06 22:18:39+00:00 | 9 | 0 |
+| 2026-05-11T22:47:11Z | 131.173.61.9 | a02609a1-120a-4f9b-b7ee-d420aa20136f | 2026-05-11 22:43:53+00:00 | 2026-05-11 22:47:08+00:00 | 3 | 0 |
+| 2026-05-12T06:27:23Z | 131.173.22.98 | 6e809242-d12c-4aad-ab47-c4f48854383a | 2026-05-12 06:23:45+00:00 | 2026-05-12 06:27:22+00:00 | 12 | 0 |
+| 2026-05-13T22:18:10Z | 131.173.24.54 | 730a4bed-1fa4-4143-98c0-bcff78c63013 | 2026-05-13 22:13:20+00:00 | 2026-05-13 22:18:01+00:00 | 9 | 0 |
+| 2026-05-18T22:44:59Z | 131.173.61.9 | a02609a1-120a-4f9b-b7ee-d420aa20136f | 2026-05-18 22:42:38+00:00 | 2026-05-18 22:44:59+00:00 | 3 | 0 |
+| 2026-05-19T06:26:10Z | 131.173.22.98 | 6e809242-d12c-4aad-ab47-c4f48854383a | 2026-05-19 06:23:19+00:00 | 2026-05-19 06:26:05+00:00 | 12 | 0 |
+| 2026-05-20T20:25:50Z | 131.173.32.4 | 3587a126-c4dc-4312-a907-d7cd933c9db5 | 2026-05-20 20:22:35+00:00 | 2026-05-20 20:25:38+00:00 | 15 | 0 |
+| 2026-05-20T22:18:21Z | 131.173.24.54 | 730a4bed-1fa4-4143-98c0-bcff78c63013 | 2026-05-20 22:13:18+00:00 | 2026-05-20 22:18:20+00:00 | 8 | 1 |
+| 2026-05-25T22:55:28Z | 131.173.61.9 | a02609a1-120a-4f9b-b7ee-d420aa20136f | 2026-05-25 22:46:01+00:00 | 2026-05-25 22:55:25+00:00 | 3 | 0 |
+| 2026-05-26T06:26:07Z | 131.173.22.98 | 6e809242-d12c-4aad-ab47-c4f48854383a | 2026-05-26 06:23:16+00:00 | 2026-05-26 06:26:02+00:00 | 12 | 0 |
+| 2026-05-27T20:24:59Z | 131.173.32.4 | 3587a126-c4dc-4312-a907-d7cd933c9db5 | 2026-05-27 20:21:44+00:00 | 2026-05-27 20:24:52+00:00 | 15 | 0 |
+| 2026-05-27T22:18:49Z | 131.173.24.54 | 730a4bed-1fa4-4143-98c0-bcff78c63013 | 2026-05-27 22:13:34+00:00 | 2026-05-27 22:18:43+00:00 | 9 | 0 |
+| de0b8bc7-9f6a-4954-b7a7-3294274eb5a8 | 131.173.22.98 | eba6af86-f06c-4cf0-979a-ac5c2795bc07 | 2026-06-17 07:25:13+00:00 | 2026-06-17 07:28:00+00:00 | 11 | 0 |
+| b063ad09-1154-42cc-9e43-e50e5a5d8301 | 131.173.32.4 | f27ab7a6-1b82-4af3-86ff-a07d520d1ea8 | 2026-06-17 07:30:25+00:00 | 2026-06-17 07:33:31+00:00 | 13 | 0 |
+| 5a457ea0-7056-49b2-a79a-19bb814d59f4 | 131.173.24.54 | ca21498c-4735-417c-a504-f45a27205173 | 2026-06-17 07:40:41+00:00 | 2026-06-17 07:45:26+00:00 | 9 | 0 |
+| artificial-report-old | 192.0.2.250 | artificial-mitigation-task | 2026-06-17 11:11:11.170769+00:00 | 2026-06-17 11:11:11.170769+00:00 | 1 | 0 |
+| artificial-report-new | 192.0.2.250 | artificial-mitigation-task | 2026-06-17 12:11:11.170769+00:00 | 2026-06-17 12:11:11.170769+00:00 | 1 | 1 |
+
+## Mitigation Events
+
+| Mitigated at report | Host | Task ID | Source report | Finding | Port/Proto | Mitigated at | Reappeared at |
+|---|---|---|---|---|---|---|---|
+| 2026-05-20T22:18:21Z | 131.173.24.54 | 730a4bed-1fa4-4143-98c0-bcff78c63013 | 2026-05-13T22:18:10Z | Services | 22/tcp | 2026-05-20 22:18:20+00:00 | 2026-05-27 22:18:43+00:00 |
+| artificial-report-new | 192.0.2.250 | artificial-mitigation-task | artificial-report-old | Artificial mitigated finding | 443/tcp | 2026-06-17 12:11:11.170769+00:00 | - |
